build_data/pyproject.toml
[project]
name = "token-embeddings-knn-build"
version = "0.1.0"
description = "Offline KNN computation for token embeddings"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "transformers",
    "safetensors",
    "huggingface-hub",
    "faiss-gpu-cu12",
    "numpy",
    "tqdm",
]
